LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-timeline-in-keyframe-expected.txt


FAIL The animation-timeline property may not be used in keyframes assert_equals: expected (string) "auto" but got (undefined) undefined
